# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Hospital Water Tank Monitoring API
  description: Backend REST API & Realtime Monitoring Specification for Integrated Hospital Water Tank Monitoring Research Project (RSA UGM).
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Local Development Server
paths:
...
  /api/alerts:
    get:
      summary: List alerts
      description: Retrieve list of threshold alerts, optionally filtered by status.
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [active, resolved]
      responses:
        '200':
          description: List of alerts
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  count:
                    type: integer
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Alert'
        '401':
          description: Unauthorized
  /api/alerts/{id}/resolve:
    patch:
      summary: Manually resolve alert (Operator / Admin)
      description: Mark an active alert as resolved. Broadcasts ALERT_RESOLVED event over WebSocket.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Alert marked as resolved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Alert'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (Operator or Admin role required)
        '404':
          description: Alert not found
    put:
      summary: Manually resolve alert (Idempotent)
      description: Mark an active alert as resolved.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Alert marked as resolved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Alert'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (Operator or Admin role required)
        '404':
          description: Alert not found
  /api/gateways:
    get:
      summary: List all LoRa gateways
      description: Returns list of Raspberry Pi gateways with their connected STM32 sensor nodes.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of gateways
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Gateway'
        '401':
          description: Unauthorized
  /api/gateways/{deviceCode}/sync:
    get:
      summary: Synchronize gateway and sensor nodes mapping
      description: Used by Raspberry Pi gateway on boot to discover monitored tank IDs, sensor node mappings, and tank capacities.
      security:
        - GatewayKeyAuth: []
        - BearerAuth: []
      parameters:
        - name: deviceCode
          in: path
          required: true
          schema:
            type: string
            example: "GW-HUB-01"
        - name: X-Gateway-Key
          in: header
          required: false
          schema:
            type: string
          description: Pre-shared gateway security key
      responses:
        '200':
          description: Gateway mappings successfully synced
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GatewaySyncResponse'
        '401':
          description: Unauthorized (Invalid gateway key or missing token)
        '404':
          description: Gateway device code not found
  /api/nodes:
    get:
      summary: List all sensor nodes
      description: Returns all STM32 sensor nodes with attached tank and gateway information.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of sensor nodes
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  count:
                    type: integer
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/SensorNode'
        '401':
          description: Unauthorized
  /api/auth/login:
    post:
      summary: User Login
      description: Authenticate user credentials and return a JWT Bearer token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginInput'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid username or password
  /api/auth/me:
    get:
      summary: Get current authenticated user
      description: Returns profile of the user currently logged in.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Profile data
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
...
  /api/reports/readings/export:
    get:
      summary: Export historical readings
      description: Download water level readings formatted as CSV.
      security:
        - BearerAuth: []
      parameters:
        - name: tankId
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: startDate
          in: query
          schema:
            type: string
            format: date-time
        - name: endDate
          in: query
          schema:
            type: string
            format: date-time
        - name: format
          in: query
          schema:
            type: string
            enum: [csv, xlsx, pdf]
            default: csv
      responses:
        '200':
          description: File download binary stream
          content:
            text/csv:
              schema:
                type: string
                format: binary
        '400':
          description: Missing required tankId
        '401':
          description: Unauthorized
  /api/reports/alerts/export:
    get:
      summary: Export alerts history
      description: Download alerts history formatted as CSV.
      security:
        - BearerAuth: []
      parameters:
        - name: level
          in: query
          schema:
            type: string
            enum: [WARNING, CRITICAL]
        - name: startDate
          in: query
          schema:
            type: string
            format: date-time
        - name: endDate
          in: query
          schema:
            type: string
            format: date-time
        - name: format
          in: query
          schema:
            type: string
            enum: [csv, xlsx, pdf]
            default: csv
      responses:
        '200':
          description: File download binary stream
          content:
            text/csv:
              schema:
                type: string
                format: binary
        '401':
          description: Unauthorized
components:
  schemas:
...
    Alert:
      type: object
      properties:
        id:
          type: integer
          example: 89
        tankId:
          type: string
          format: uuid
        tankName:
          type: string
          example: "Tangki Gedung IBS"
        tankLocation:
          type: string
          nullable: true
        level:
          type: string
          enum: [WARNING, CRITICAL]
          example: "CRITICAL"
        message:
          type: string
          example: "Level air tangki IBS CRITICAL: 24.5%"
        status:
          type: string
          enum: [active, resolved]
          example: "active"
        createdAt:
          type: string
          format: date-time
        resolvedAt:
          type: string
          format: date-time
          nullable: true
    Gateway:
      type: object
      properties:
        id:
          type: string
          format: uuid
        deviceCode:
          type: string
          example: "GW-HUB-01"
        status:
          type: string
          example: "online"
        lastHeartbeat:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        nodesCount:
          type: integer
          example: 3
        nodes:
          type: array
          items:
            $ref: '#/components/schemas/SensorNode'
    GatewaySyncResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        gatewayCode:
          type: string
          example: "GW-HUB-01"
        syncedAt:
          type: string
          format: date-time
        mappings:
          type: array
          items:
            type: object
            properties:
              deviceCode:
                type: string
                example: "STM32-GEDUNG-A"
              sensorNodeId:
                type: string
                format: uuid
              tankId:
                type: string
                format: uuid
              tankName:
                type: string
                example: "Tangki Atap Gedung Utama"
              capacityLiters:
                type: string
                example: "10000.00"
    SensorNode:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tankId:
          type: string
          format: uuid
        gatewayId:
          type: string
          format: uuid
          nullable: true
        deviceCode:
          type: string
          example: "STM32-GEDUNG-A"
        sensorType:
          type: string
          example: "ultrasonic"
        status:
          type: string
          example: "active"
        lastRssi:
          type: integer
          nullable: true
          example: -78
        lastSeen:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
          example: "admin"
        role:
          type: string
          enum: [admin, operator, viewer]
          example: "admin"
        isActive:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
    LoginInput:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          example: "admin"
        password:
          type: string
          example: "password123"
    LoginResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Login berhasil"
        data:
          type: object
          properties:
            token:
              type: string
              example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
            user:
              $ref: '#/components/schemas/User'
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Enter JWT Bearer token obtained from /api/auth/login
    GatewayKeyAuth:
      type: apiKey
      in: header
      name: X-Gateway-Key
      description: Pre-shared API Key header for LoRa Gateway synchronization